Move RemindFeatureUtil to the command.flashcard package.
