+++ src/main/java/seedu/address/logic/commands/flashcard/RemindCommand.java
@@ -10,7 +10,6 @@ import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.commandresults.FlashcardCommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.commands.RemindFeatureUtil;
 import seedu.address.model.Model;
 import seedu.address.model.flashcard.Flashcard;
