+++ src/main/java/seedu/address/logic/commands/flashcard/RemindFeatureUtil.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.flashcard;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
