+++ src/main/java/seedu/address/logic/commands/global/ExitCommand.java
@@ -8,7 +8,7 @@ import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandHistory;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.commandresults.GlobalCommandResult;
-import seedu.address.logic.commands.RemindFeatureUtil;
+import seedu.address.logic.commands.flashcard.RemindFeatureUtil;
 import seedu.address.model.Model;
 import seedu.address.model.flashcard.Flashcard;
